[HACK] provide HBase source support for cascalog
This HACK enables using an HBase table as a source in a cascalog flow.
It does this by stripping the cascalog magical characters ('?', '!')
from the beginning of field names. This behavior is enabled via a new
configuration point, cascading.hbase.cascalogsafe.

This is an incomplete implementation. The HBaseScheme is missing
important functionality like the concepts of strictness and safety
provided by the TextDelimited scheme.

An example of consuming this scheme from cascalog is provided in
https://gist.github.com/4516811